remove auth0
