+++ front/lib/iam/provider.ts
@@ -4,7 +4,7 @@ export interface ExternalUser {
   name: string;
   nickname: string;
   auth0Sub: string | null;
-  workOSUserId: string | null;
+  workOSUserId: string;
   // Google-specific fields.
   family_name?: string;
   given_name?: string;
@@ -14,7 +14,7 @@ export interface ExternalUser {
 }
 
 export type SessionWithUser = {
-  type: "workos" | "auth0";
+  type: "workos";
   sessionId: string;
   user: ExternalUser;
   workspaceId?: string;
